refactor: restore owner-crate boundaries for node adapters
Move BIP9/MTP rules to consensus, tree walks and BlockBodySource to
chain, mining candidate context to mining, P2P serving to p2p, and
txindex status projection to RPC. Node now only wires those surfaces:
a private IndexBlockSource, gateway observer legs, and composition.

bitcoin.conf stays in the binary input layer; the unused UtxoSetView
wrapper is deleted rather than relocated.
