+++ crates/node/src/txindex_worker_query_tests.rs
@@ -324,7 +324,7 @@ impl QueryFixture {
             let source: Arc<dyn BlockBodySource> = source.clone();
             source
         });
-        let block_source = NodeBlockSource::new(Arc::new(RwLock::new(
+        let block_source = IndexBlockSource::new(Arc::new(RwLock::new(
             records
                 .into_iter()
                 .collect::<bitcoin_rs_rpc::context::BlockLog>(),
