+++ docs/contracts/indexing.md
@@ -7,8 +7,8 @@ rebuilds.
 Owners:
 - `TxIndexRuntime`, `TxIndexQueryEngine`, `Worker` in `crates/node/src/txindex_worker.rs`
 - `IndexWriter`, `IndexReader`, `IndexCapabilities`, `IndexCapability`, `IndexWatermarks`, `IndexWatermark` in `crates/index/src/index.rs` and `crates/index/src/types.rs`
-- Capability status provider in `crates/node/src/capabilities.rs` and
-  `crates/rpc/src/context.rs`
+- Capability status: worker-owned facts in `crates/node/src/txindex_worker.rs`
+  and RPC projection in `crates/rpc/src/capabilities.rs`
 
 ## Clauses
 
@@ -165,5 +165,7 @@ remove another script's output.
   publication, open failure/timeout, and shutdown abandonment.
 - `crates/node/src/txindex_worker_query_tests.rs`: query gating, snapshot
   consistency, and revision ABA detection tests.
+- `crates/node/src/txindex_worker_block_source_tests.rs`: confirmed-body
+  serving by height/hash (`IDX-03`, `RCV-01`).
 - `crates/node/src/apply.rs`:
   `txindex_worker_failure_makes_queries_unavailable_without_blocking_apply`.
