+++ docs/policies/p2p-compatibility.md
@@ -4,7 +4,7 @@ This document declares bitcoin-rs's peer-wire compatibility contract with Bitcoi
 
 ## 1. Scope and Authority
 
-This policy applies to the P2P transport and peer-protocol surface of `bitcoin-rs-p2p` (`crates/p2p`), the P2P chain-serving adapter (`crates/node/src/p2p_chain.rs`), and the network flags of the node binary. It is the peer-visible counterpart to `docs/policies/source-compatibility.md` (toolchain) and the RPC compatibility manifest (track 4a). Where this document and prose comments disagree, this document wins; where it and the code disagree, the code is the defect.
+This policy applies to the P2P transport and peer-protocol surface of `bitcoin-rs-p2p` (`crates/p2p`), including `ActiveChainQuery` (`crates/p2p/src/chain_query.rs`), and the network flags of the node binary. It is the peer-visible counterpart to `docs/policies/source-compatibility.md` (toolchain) and the RPC compatibility manifest (track 4a). Where this document and prose comments disagree, this document wins; where it and the code disagree, the code is the defect.
 
 ## 2. Pinned Reference Version
 
@@ -128,6 +128,6 @@ Explicit deltas from Core 31.1, each intentional and safe:
 - **Deterministic fixtures**: `crates/p2p/tests/core_compat.rs` pins the handshake fields and service bits, per-network magic/ports and framing, getheaders/headers semantics and bounds, inv/getdata relay round-trips with `notfound`, the reject-or-ignore matrix of §6, and the peer-visible behavior across a chain switch (reorg) and a restart at the `ChainQuery` seam: a rebuilt query serves byte-identical answers, a switched active branch serves the new branch from the fork point and `notfound`s stale bodies. Run with `cargo test -p bitcoin-rs-p2p`.
 - **Fuzz**: `fuzz/fuzz_targets/p2p_message.rs` drives all 36 payload decoders; its `COMMANDS` list must stay in step with `decode_payload` (a missing entry is a decoder no fuzz input can reach).
 - **Live lane (cut, env-gated)**: `scripts/run-p2p-core-interop.sh --bitcoind-command <cmd>` drives a real Bitcoin Core 31.x (regtest) plus a bitcoin-rs node through the initial sync, mines extra blocks after the handshake to prove the node follows Core's announcements while connected (bitcoin-rs itself sends no proactive announcements; see the deviation ledger), records Core's own `getpeerinfo` view of us (services bits, subver) into an evidence JSON, and runs the `#[ignore]`d verifier `crates/p2p/tests/core_interop_live.rs`. The lane is never run in CI (no bitcoind on CI hosts); its evidence belongs under `docs/benchmarks/` when a Core bump is pinned.
-- Node-level reorg is implemented: `crates/node/src/reorg.rs` (`switch_to_branch`, `invalidate_block`) moves the applied tip off a losing branch, and sync calls `switch_to_branch` when a higher-work header branch wins (`crates/node/src/sync.rs`). The reorg fixture pins the peer-visible part of this at the `ChainQuery` seam — the exact surface `NodeP2pChainQuery` implements — via `reorg_switches_which_chain_a_peer_sees` (`crates/p2p/tests/core_compat.rs`).
+- Node-level reorg is implemented: `crates/node/src/reorg.rs` (`switch_to_branch`, `invalidate_block`) moves the applied tip off a losing branch, and sync calls `switch_to_branch` when a higher-work header branch wins (`crates/node/src/sync.rs`). The reorg fixture pins the peer-visible part of this at the `ChainQuery` seam — the exact surface `ActiveChainQuery` implements — via `reorg_switches_which_chain_a_peer_sees` (`crates/p2p/tests/core_compat.rs`).
 
 See also [docs/contracts/p2p-wire.md](../contracts/p2p-wire.md) for the contracts index and precedence rule.
